requirements.txt
Flask
google-genai
python-dotenv
gunicorn